+++ sql/003_permis.sql
@@ -1,8 +1,8 @@
 DROP TABLE IF EXISTS permis;
 CREATE TABLE permis (
-	id INTEGER(9) NOT NULL,
-	nom VARCHAR(100) NOT NULL,
-	PRIMARY KEY (id)
+    id INT NOT NULL,
+    nom VARCHAR(100) NOT NULL,
+    PRIMARY KEY (id)
 ) TYPE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
 
 INSERT INTO permis VALUES (1,'Permis A');
